changing name test, adding navigation to screen and moving files in generic folder
