+++ app/src/androidTest/java/com/github/se/polyfit/ui/screen/HomeScreen.kt
@@ -6,8 +6,8 @@ import io.github.kakaocup.compose.node.element.KNode
 
 class HomeScreen(semanticsProvider: SemanticsNodeInteractionsProvider) :
     ComposeScreen<HomeScreen>(
-        semanticsProvider = semanticsProvider, viewBuilderAction = { hasTestTag("LoginScreen") }) {
+        semanticsProvider = semanticsProvider, viewBuilderAction = { hasTestTag("HomeScreen") }) {
 
   // Structural elements of the UI
-  val homeScreen: KNode = child { hasTestTag("HomeScreen") }
+  val homeScreen: KNode = child { hasTestTag("HomeScreenText") }
 }
