+++ app/src/androidTest/java/com/github/se/polyfit/ui/screen/LoginScreen.kt
@@ -9,7 +9,6 @@ class LoginScreen(semanticsProvider: SemanticsNodeInteractionsProvider) :
         semanticsProvider = semanticsProvider, viewBuilderAction = { hasTestTag("LoginScreen") }) {
 
   // Structural elements of the UI
-  val loginScreen: KNode = child { hasTestTag("LoginScreen") }
   val loginTitle: KNode = child { hasTestTag("LoginTitle") }
   val loginButton: KNode = child { hasTestTag("LoginButton") }
 }
