+++ app/src/main/java/com/github/se/polyfit/MainActivity.kt
@@ -26,7 +26,7 @@ class MainActivity : ComponentActivity() {
         val navigation = Navigation(navController)
         NavHost(navController = navigation.getController(), startDestination = Route.Register) {
           composable(Route.Register) { LoginScreen(navigation::navigateToHome) }
-          composable(Route.Home) { HomeScreen() }
+          composable(Route.Home) { HomeScreen(navigation) }
         }
       }
     }
