+++ app/src/main/java/com/github/se/polyfit/ui/screen/HomeScreen.kt
@@ -3,8 +3,9 @@ package com.github.se.polyfit.ui.screen
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.platform.testTag
+import com.github.se.polyfit.ui.navigation.Navigation
 
 @Composable
-fun HomeScreen() {
-  Text("Home Screen", modifier = androidx.compose.ui.Modifier.testTag("HomeScreen"))
+fun HomeScreen(navigation: Navigation) {
+  Text("Home Screen", modifier = androidx.compose.ui.Modifier.testTag("HomeScreenText"))
 }
